split out kanwa dict generation
