+++ setup.py
@@ -4,116 +4,27 @@ import io
 import os
 import re
 
-from klepto.archives import file_archive
 from setuptools import setup
 from setuptools.command.build_py import build_py
-from six import unichr
+
+from kakasidict import Genkanwadict
 
 package_name = "pykakasi"
 root_dir = os.path.abspath(os.path.dirname(__file__))
 
 
-class genkanwadict(object):
-    records = {}
-
-    def run(self, src, dst):
-        with open(src, 'rb') as f:
-            for line in f:
-                self.parsekdict(line.decode("utf-8"))
-            f.close()
-        self.kanwaout(dst)
-
-    # for itaiji and kana dict
-
-    def mkdict(self, src, dst):
-        max_key_len = 0
-        dic = {}
-        with open(src, "rb") as f:
-            for line in f:
-                line = line.decode("utf-8").strip()
-                if line.startswith(';;'):  # skip comment
-                    continue
-                if re.match(r"^$", line):
-                    continue
-                try:
-                    (v, k) = (re.sub(r'\\u([0-9a-fA-F]{4})',
-                                     lambda x: unichr(int(x.group(1), 16)), line)).split(' ')
-                    dic[k] = v
-                    max_key_len = max(max_key_len, len(k))
-                except ValueError:
-                    raise Exception("Cannot process dictionary line: ", line)
-        d = file_archive(dst, dic, serialized=True)
-        d['_max_key_len_'] = max_key_len
-        d.dump()
-
-    # for kanwadict
-
-    def parsekdict(self, line):
-        line = line.strip()
-        if line.startswith(';;'):  # skip comment
-            return
-        (yomi, kanji) = line.split(' ')
-        if ord(yomi[-1:]) <= ord('z'):
-            tail = yomi[-1:]
-            yomi = yomi[:-1]
-        else:
-            tail = ''
-        self.updaterec(kanji, yomi, tail)
-
-    def updaterec(self, kanji, yomi, tail):
-        key = "%04x" % ord(kanji[0])
-        if key in self.records:
-            if kanji in self.records[key]:
-                rec = self.records[key][kanji]
-                rec.append((yomi, tail))
-                self.records[key].update({kanji: rec})
-            else:
-                self.records[key][kanji] = [(yomi, tail)]
-        else:
-            self.records[key] = {}
-            self.records[key][kanji] = [(yomi, tail)]
-
-    def kanwaout(self, out):
-        dic = file_archive(out, self.records, serialized=True)
-        dic.dump()
-
-
 class MyBuild(build_py):
-
-    def generate_dictionaries(self):
-        DICTS = [
-            ('itaijidict.utf8', 'itaijidict3.db'),
-            ('hepburndict.utf8', 'hepburndict3.db'),
-            ('kunreidict.utf8', 'kunreidict3.db'),
-            ('passportdict.utf8', 'passportdict3.db'),
-            ('hepburnhira.utf8', 'hepburnhira3.db'),
-            ('kunreihira.utf8', 'kunreihira3.db'),
-            ('passporthira.utf8', 'passporthira3.db')
-        ]
-        kanwa = genkanwadict()
-        dstdir = os.path.join(self.build_lib, 'pykakasi', 'data')
-        if not os.path.exists(dstdir):
-            os.mkdir(dstdir)
-        for (src_f, pkl_f) in DICTS:
-            src = os.path.join('src', 'data', src_f)
-            dst = os.path.join(self.build_lib, 'pykakasi', 'data', pkl_f)
-            if (os.path.exists(dst)):
-                os.unlink(dst)
-            kanwa.mkdict(src, dst)
-        src = os.path.join('src', 'data', 'kakasidict.utf8')
-        dst = os.path.join(self.build_lib, 'pykakasi', 'data', 'kanwadict4.db')
-        if (os.path.exists(dst)):
-            os.unlink(dst)
-        kanwa.run(src, dst)
 
     def run(self):
         build_py.run(self)
         if not self.dry_run:
-            self.generate_dictionaries()
+            kanwa = Genkanwadict()
+            dstdir = os.path.join(self.build_lib, 'pykakasi', 'data')
+            kanwa.generate_dictionaries(dstdir)
 
 
 def readme():
-    with io.open(os.path.join(os.path.dirname(__file__), 'README.rst'), mode="r", encoding="UTF-8") as f:
+    with io.open(os.path.join(root_dir, 'README.rst'), mode="r", encoding="UTF-8") as f:
         return f.read()
