+++ src/pykakasi/__init__.py
@@ -1,5 +1,6 @@
 from .kakasi import kakasi, wakati
 from .kanji import J2
+from .properties import Configurations
 from .scripts import A2, H2, K2, Sym2
 
 __copyright__ = 'Copyright (C) 2010-2019 Hiroshi Miura'
@@ -9,4 +10,4 @@ __author__ = 'Hiroshi Miura'
 __author_email__ = 'miurahr@linux.com'
 __url__ = 'http://github.com/miurahr/pykakasi'
 
-__all__ = ["kakasi", "wakati", "exceptions", "J2", "K2", "H2", "Sym2", "A2"]
+__all__ = ["kakasi", "wakati", "exceptions", "J2", "K2", "H2", "Sym2", "A2", "Configurations"]
